+++ src/index.js
@@ -3,8 +3,8 @@
 //@spencermountain
 const fetch = require('./lib/fetch_text');
 const parse = require('./parse');
-const markdown = require('./markdown');
-const html = require('./html');
+const markdown = require('./output/markdown');
+const html = require('./output/html');
 const version = require('../package').version;
 
 //use a global var for lazy customization
