+++ src/output/markdown/infobox.js
