+++ src/parse/section/list.js
@@ -2,7 +2,7 @@ const list_reg = /^[#\*:;\|]+/;
 const bullet_reg = /^\*+[^:,\|]{4}/;
 const number_reg = /^ ?\#[^:,\|]{4}/;
 const has_word = /[a-z]/i;
-const parseLine = require('./sentence/line');
+const parseLine = require('./sentence/').parseLine;
 
 // does it start with a bullet point or something?
 const isList = function(line) {
