+++ src/parse/section/sentence/templates/index.js
@@ -1,0 +1,12 @@
+// const parsers = {
+//   main: (section) => {
+//
+//   }
+// };
+//
+const parseTemplates = function(obj) {
+  // let names = obj.text.match(/^\{\{([^:|\n ]+)/g);
+  // console.log(names);
+  return obj;
+};
+module.exports = parseTemplates;
